+++ src/Query/Builder.php
@@ -531,24 +531,6 @@ class Builder extends BaseBuilder
     }
 
     /**
-     * Add a "where all" clause to the query.
-     *
-     * @param  string  $column
-     * @param  array  $values
-     * @param  string  $boolean
-     * @param  bool  $not
-     * @return $this
-     */
-    public function whereAll($column, array $values, $boolean = 'and', $not = false)
-    {
-        $type = 'all';
-
-        $this->wheres[] = compact('column', 'type', 'boolean', 'values', 'not');
-
-        return $this;
-    }
-
-    /**
      * @inheritdoc
      * @param list{mixed, mixed}|CarbonPeriod $values
      */
@@ -1042,17 +1024,6 @@ class Builder extends BaseBuilder
         }
 
         return $compiled;
-    }
-
-    /**
-     * @param  array  $where
-     * @return array
-     */
-    protected function compileWhereAll(array $where): array
-    {
-        extract($where);
-
-        return [$column => ['$all' => array_values($values)]];
     }
 
     /**